requirements.txt
streamlit
numpy
opencv-python-headless
pandas
pillow
tensorflow
keras